# docs/v25-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Predator Analytics API
  description: |
    # 🛡️ Predator Analytics v25.0 API

    Self-healing analytics platform with multi-agent AI orchestration.

    ## Основні можливості

    - **Cases Management** — Управління справами для аналізу
    - **Data Hub** — Завантаження та обробка даних
    - **AI Agents** — Мультиагентна система аналітики
    - **Real-time Analytics** — Аналітика в реальному часі

    ## Автентифікація

    API використовує JWT Bearer tokens. Отримайте токен через `/auth/login`.

  version: "25.0.0"
  contact:
    name: Predator Team
    email: support@predator.ai
  license:
    name: Proprietary
servers:
  - url: http://localhost:8090
    description: Local Development
  - url: https://api.predator.ai
    description: Production
...
paths:
...
  /api/v25/data/sources:
    get:
      tags: [Data]
      summary: List data sources
      description: Get available data sources and their status
      operationId: listDataSources
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of data sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataSource"
  /api/v25/data/upload:
    post:
      tags: [Data]
      summary: Upload data file
      description: Upload a file for processing (CSV, Excel, JSON)
      operationId: uploadData
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
                source_name:
                  type: string
                  description: Name for the data source
                auto_process:
                  type: boolean
                  default: true
                  description: Automatically start ETL processing
      responses:
        "202":
          description: Upload accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DataSource:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [file, database, api, stream]
        status:
          type: string
          enum: [active, inactive, error]
        records_count:
          type: integer
        last_updated:
          type: string
          format: date-time
    UploadResponse:
      type: object
      properties:
        file_id:
          type: string
          format: uuid
        filename:
          type: string
        size_bytes:
          type: integer
        status:
          type: string
        etl_job_id:
          type: string
          format: uuid
...